+++ resolvers/event.js
@@ -6,9 +6,9 @@ const { AuthenticationError } = require('apollo-server');
 const eventResolvers = {
   Query: {
     events: async (obj, args, context) => {
-      let hasPerm = await context.group.hasPerm("view_event");
+      let hasPerm = await context.group.hasPerm(context.group.id, "read_event");
       if (!hasPerm) {
-          return null;
+        throw new AuthenticationError("Unauthorized. Event not read.");
       }
       return db.event.findAll();
     },
@@ -18,8 +18,13 @@ const eventResolvers = {
     createdBy: (obj, args, context, info) => db.user.findByPk(obj.createdBy)
   },
   Mutation: {
-    addEvent: async (parent, args) => {
+    addEvent: async (parent, args, context) => {
       // Check if createdBy is valid
+
+      let hasPerm = await context.group.hasPerm(context.group.id, "add_event");
+      if (!hasPerm) {
+        throw new AuthenticationError("Unauthorized. Event not created.");
+      }
       const user = await db.user.findByPk(args.createdBy);
       if (!user) {
         throw new Error("Invalid user ID");
@@ -31,7 +36,12 @@ const eventResolvers = {
         isActive: args.isActive
       });
     },
-    updateEvent: async (parent, args) => {
+    updateEvent: async (parent, args, context) => {
+
+      let hasPerm = await context.group.hasPerm(context.group.id, "update_event");
+      if (!hasPerm) {
+        throw new AuthenticationError("Unauthorized. Event not updated.");
+      }
       const event = await db.event.findByPk(args.id);
       if (args.createdBy) {
         const user = await db.user.findByPk(args.createdBy);
@@ -43,7 +53,7 @@ const eventResolvers = {
         throw new Error("Invalid event ID");
       }
 
-      await db.event.update(
+      await db.event.update(  
         {
           name: args.name,
           eventDate: args.eventDate,
@@ -56,9 +66,15 @@ const eventResolvers = {
       );
       return db.event.findByPk(args.id);
     },
-    deleteEvent: (parent, args) => {
+
+    deleteEvent: (parent, args, context) => {
       // Return status for destroy
       // 1 for successful deletion, 0 otherwise
+      let hasPerm = context.group.hasPerm(context.group.id, "delete_event");
+      if (!hasPerm) {
+        throw new AuthenticationError("Unauthorized. Event not deleted.");
+      
+      }
       return db.event.destroy({
         where: {
           id: args.id
